refactor: move use module to modules
