+++ src/auth/models/AuthRequest.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
-import { User } from '../../user/entities/user.entity';
+import { User } from 'src/modules/user/entities/user.entity';
 
 export interface AuthRequest extends Request {
   user: User;
